feat: useModal 훅 생성
