+++ src/hooks/useModal.tsx
@@ -1,0 +1,21 @@
+import { useState } from 'react';
+
+const useModal = () => {
+  const [modalOpen, setModalOpen] = useState(false);
+
+  const handleModalClose = () => {
+    setModalOpen(false);
+  };
+
+  const handleModalOpen = () => {
+    setModalOpen(true);
+  };
+
+  return {
+    modalOpen,
+    handleModalOpen,
+    handleModalClose,
+  };
+};
+
+export default useModal;
